refactor : 노선 관리 Feature
예외가 발생하면 Controller 에서 에러가 담긴 Response 를 리턴하도록 구현

+++ src/main/java/nextstep/subway/applicaion/LineService.java
@@ -5,11 +5,13 @@ import nextstep.subway.applicaion.dto.LineRequest;
 import nextstep.subway.applicaion.dto.LineResponse;
 import nextstep.subway.domain.Line;
 import nextstep.subway.domain.LineRepository;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -32,26 +34,39 @@ public class LineService {
         );
     }
 
+    @Transactional(readOnly = true)
     public List<LineDetailResponse> getLineList() {
         return lineRepository.findAll().stream()
                 .map(LineDetailResponse::from)
                 .collect(Collectors.toList());
     }
 
+    @Transactional(readOnly = true)
     public LineDetailResponse getLine(Long id) {
-        return LineDetailResponse.from(lineRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("[ERROR] 존재하지 않는 노선입니다.")
-                ));
+        Optional<Line> optionalLine = lineRepository.findById(id);
+
+        if (optionalLine.isPresent()) {
+            return LineDetailResponse.from(optionalLine.get());
+        }
+
+        return null;
     }
 
-    public void patchLine(Long id, LineRequest lineRequest) {
-        Line line = lineRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("[ERROR] 존재하지 않는 노선입니다."));
+    public boolean patchLine(Long id, LineRequest lineRequest) {
+        Optional<Line> optionalLine = lineRepository.findById(id);
 
-        line.modify(lineRequest.getName(), lineRequest.getColor());
+        if (optionalLine.isPresent()) {
+            optionalLine.get().modify(lineRequest.getName(), lineRequest.getColor());
+            return false;
+        }
+        return true;
     }
 
-    public void deleteLine(Long id) {
-        lineRepository.deleteById(id);
+    public boolean deleteLine(Long id) {
+        if (lineRepository.existsById(id)) {
+            lineRepository.deleteById(id);
+            return false;
+        }
+        return true;
     }
 }
